refactor: added pick for omitting extra args for hooks
