+++ packages/core/src/utils/index.ts
@@ -1,1 +1,2 @@
+export { pick } from './pick';
 export { warn } from './warn';
