+++ packages/react/src/hooks/stream/useStream.ts
@@ -1,7 +1,17 @@
-import { GetStreamArgs, LivepeerProvider, Stream, getStream } from 'livepeer';
+import {
+  GetStreamArgs,
+  LivepeerProvider,
+  Stream,
+  getStream,
+  pick,
+} from 'livepeer';
 
 import { QueryClientContext } from '../../context';
-import { UseInternalQueryOptions, useInternalQuery } from '../../utils';
+import {
+  UseInternalQueryOptions,
+  useInternalQuery,
+  useInternalQueryKeys,
+} from '../../utils';
 import { useLivepeerProvider } from '../providers';
 
 export function useStream<TLivepeerProvider extends LivepeerProvider>(
@@ -14,6 +24,6 @@ export function useStream<TLivepeerProvider extends LivepeerProvider>(
     queryKey: [{ entity: 'getStream', args, livepeerProvider }],
     queryFn: async () => getStream<TLivepeerProvider>(args as GetStreamArgs),
     enabled: Boolean(typeof args === 'string' ? args : args?.streamId),
-    ...(typeof args === 'object' ? args : {}),
+    ...(typeof args === 'object' ? pick(args, useInternalQueryKeys) : {}),
   });
 }
