+++ packages/react/src/utils/index.ts
@@ -1,7 +1,9 @@
 export { deserialize } from './deserialize';
 export {
   useInternalMutation,
+  useInternalMutationKeys,
   useInternalQuery,
+  useInternalQueryKeys,
   type Status,
   type UseInternalMutationOptions,
   type UseInternalQueryOptions,
