+++ packages/react/src/utils/query/index.ts
@@ -1,9 +1,11 @@
 export type { Status } from './types';
 export {
   useInternalMutation,
+  useInternalMutationKeys,
   type UseInternalMutationOptions,
 } from './useInternalMutation';
 export {
   useInternalQuery,
+  useInternalQueryKeys,
   type UseInternalQueryOptions,
 } from './useInternalQuery';
